[BI-695] - Refactoring "views" folder structure
