+++ src/views/account/AccountCreationSuccess.vue
