+++ src/views/program/ProgramUsersManagement.vue
@@ -1,6 +1,5 @@
 <!--
-  - See the NOTICE file distributed with this work for additional information
-  - regarding copyright ownership.
+  - See the NOTICE file distributed with this work for additional information regarding copyright ownership.
   -
   - Licensed under the Apache License, Version 2.0 (the "License");
   - you may not use this file except in compliance with the License.
@@ -25,7 +24,7 @@
 
 <script lang="ts">
   import { Component, Prop, Vue } from 'vue-property-decorator'
-  import ProgramUsersTable from "../components/program/ProgramUsersTable.vue";
+  import ProgramUsersTable from "../../components/program/ProgramUsersTable.vue";
   import ProgramsBase from "@/components/program/ProgramsBase.vue";
 
   @Component({
