+++ src/views/trait/Traits.vue
