+++ src/views/trait/TraitsArchived.vue
