+++ src/views/trait/TraitsFavorites.vue
